+++ Assets/Scripts/CoinData.cs
@@ -1,0 +1,16 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.IO;
+using UnityEngine;
+
+public class CoinData
+{
+    public int playerCoins;
+
+
+    public CoinData(int playerCoins = 0)
+    {
+        this.playerCoins = playerCoins;
+    }
+}