+++ Assets/Scripts/PaddleSkin.cs
@@ -1,0 +1,15 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PaddleSkin : MonoBehaviour
+{
+    /* Properties of a paddle skin, from its name to its icon */
+    public int id;
+    
+    public string displayName;
+
+    public Sprite iconSprite, paddleSprite;
+
+    public int coinPrice;
+}