+++ Assets/Scripts/managers/MainMenu.cs
@@ -11,16 +11,22 @@ public class MainMenu : MonoBehaviour
     /* Handles the main menu loop */
 
     private HelpMenuUI _helpMenuUI;
+    private ShopUI _shopUI;
+    private CoinsUI _coinsUI;
 
 
     void Awake()
     {
         _helpMenuUI = GameObject.FindGameObjectWithTag("HelpMenuUI").GetComponent<HelpMenuUI>();
+        _shopUI = GameObject.FindGameObjectWithTag("ShopUI").GetComponent<ShopUI>();
+        _coinsUI = GameObject.FindGameObjectWithTag("CoinUI").GetComponent<CoinsUI>();
     }
 
     void Start()
     {
         _helpMenuUI.Hide();
+        _shopUI.Hide();
+        _coinsUI.Hide();
     }
 
     // Enters the game via a new scene
@@ -29,6 +35,12 @@ public class MainMenu : MonoBehaviour
         SceneManager.LoadScene("SampleScene");
     }
     
+    // Opens the shop menu
+    public void OpenShop()
+    {
+        _shopUI.Show();
+    }
+
     // Quits the game
     public void QuitGame()
     {