+++ CHANGELOG.md
@@ -6,6 +6,15 @@ Written by fffadedev.
 
 ## [Unreleased]
 
+## [v1.2.0] Alpha, 2023-12-08
+
+### Added
+
+- Earning coins at the end of the game based on your performance
+- New shop to buy skins for your paddle!
+- Auto-saving of skins and coins when exiting the game
+- Shop is available during pause
+
 ## [v1.1.1] Alpha, 2023-12-04
 
 ### Fixed
@@ -26,5 +35,6 @@ Written by fffadedev.
 
 - [Back to top](#changelog)
 - [unreleased](#unreleased)
+- [1.2.0](#v1.2.0)
 - [1.1.1](#v1.1.1)
 - [1.1.0](#v1.1.0)